+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project (partially) adheres to [Semantic Versioning](https://semver.org
 ## [Unreleased]
 ### Added
 - Python 3.13 & 3.14 official support
+- `entries_color` config option validation
 
 ## [v4.15.0.0] - 2024-09-30
 ### Added